requirements.txt
tensorflow-cpu==2.20.0
pandas 
numpy 
scikit-learn
tensorboard
matplotlib
streamlit
scikeras
ipykernel
streamlit
nltk